pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pocketteam"
version = "0.1.0"
description = "Your autonomous AI IT team. 11 agents. 10 safety layers. Zero API costs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["ai-agent", "agent-framework", "autonomous-agents", "multi-agent", "claude-code", "agentic-ai", "self-healing", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "claude-agent-sdk>=0.0.9",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    "python-telegram-bot>=21.0",
    "watchdog>=4.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML",
]

[project.scripts]
pocketteam = "pocketteam.cli:main"

[project.urls]
Homepage = "https://github.com/Farid046/pocketteamtest"
Repository = "https://github.com/Farid046/pocketteamtest"
Issues = "https://github.com/Farid046/pocketteamtest/issues"

[tool.hatch.build.targets.wheel]
packages = ["pocketteam"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
